% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances_data.R
\docType{data}
\name{distances_data}
\alias{distances_data}
\title{Gravity distance data}
\format{data table with computed distances}
\usage{
distances_data
}
\description{
Data from Hinz (2018)
}
\keyword{datasets}
