% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_data}
\alias{get_data}
\title{Get additional distances datasets}
\usage{
get_data(data, years, data_url, data_store)
}
\arguments{
\item{data}{Character string containing name of dataset}

\item{years}{Vector of years for which data is requested (character)}

\item{data_url}{Character string indicating url from which data should be downloaded.}

\item{data_store}{Logical indicator whether downloaded data should be stored permanently.}
}
\description{
Provides distances data from additional available downloadable datasets
}
