% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remove_data}
\alias{remove_data}
\title{Remove some or all downloaded additional distances datasets}
\usage{
remove_data(data = NULL)
}
\arguments{
\item{data}{Character string or vector containing name of dataset(s)}
}
\description{
Remove some or all downloaded additional distances datasets
}
\examples{
remove_data("distances_from_canada_provinces_to_canada_provinces")

}
